requirements.txt
pytest
pylint
flake8
mypy
pytest-pylint
pytest-flake8
pytest-mypy
biopython
requests
